refactor: frame 패키지 분리
